Fully automate dev setup with Gitpod (#5)
* Update blank.yml

* Update elsearch.sh

* Update blank.yml

* Update elsearch.sh

* Update elsearch.sh

* Update elsearch.sh

* Delete main.yml

* List running elasticsearch processes

* Update elsearch.sh

List running processes

* Update elsearch.sh

Enable communication ports 0-400 via tcp

* Set up CI with Azure Pipelines

[skip ci]

* Fully automate dev setup with Gitpod

This commit implements a fully-automated development setup using Gitpod.io, an
online IDE for GitHub and GitLab that enables Dev-Environments-As-Code.
This makes it easy for anyone to get a ready-to-code workspace for any branch,
issue or pull request almost instantly with a single click.